+++ scripts/build-httpd.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
+# Any subsequent(*) commands which fail will cause the shell script to exit immediately
+set -e
 
 # Build and install apache
 echo "Compiling Apache..."