+++ scripts/get-httpd.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
+# Any subsequent(*) commands which fail will cause the shell script to exit immediately
+set -e
 
 # Check for Httpd Source files
 echo "Checking for Httpd Source Files..."